Fix bad Task methed call

+++ src/net/sf/memoranda/TaskList.java
@@ -20,7 +20,7 @@ public interface TaskList {
     Task getTask(String id);
 
     Task createTask(CalendarDate startDate, CalendarDate endDate, String text, int priority, long effort, long actualEffort, long timestamp,
-    		String description, String parentTaskId, long estLOC, long actLOC);
+    		String description, String parentTaskId, int estLOC, int actLOC);
 
     void removeTask(Task task);
